Increase max package size to ~8GB (#100)
* Increase max package size to ~8GB

* Add comment for max package size

* Make max package size configurable

* Add docker hub link to docs

* Also use configurable max package size for IIS delpoyments.
