+++ src/BaGetter/appsettings.json
@@ -2,6 +2,7 @@
   "ApiKey": "",
   "PackageDeletionBehavior": "Unlist",
   "AllowPackageOverwrites": false,
+  "MaxPackageSizeGiB": 8,
 
   "Database": {
     "Type": "Sqlite",
